backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-environ==0.12.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.1.0
django-extensions==3.2.3
faker==20.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
drf-yasg==1.21.7
django-ratelimit==4.1.0
pyotp==2.9.0
qrcode==7.4.2
pytz==2023.3
django-celery-beat==2.5.0
stripe==7.0.0
sendgrid==6.11.0
django-redis==5.4.0
